Separating components into separate files
